VRControllerExample.kt. VRControllerAdvancedExample.kt: revive simple VRController example and clean up and extend documentation on both examples
relates to #514
