+++ src/test/kotlin/graphics/scenery/tests/examples/advanced/VRControllerExample.kt
@@ -1,55 +1,36 @@
 package graphics.scenery.tests.examples.advanced
 
+import org.joml.Vector3f
 import graphics.scenery.*
-import graphics.scenery.attribute.material.Material
 import graphics.scenery.backends.Renderer
 import graphics.scenery.controls.OpenVRHMD
 import graphics.scenery.controls.TrackedDeviceType
 import graphics.scenery.controls.TrackerRole
-import graphics.scenery.controls.behaviours.*
-import graphics.scenery.controls.behaviours.VRSelectionWheel.Companion.toActions
 import graphics.scenery.numerics.Random
-import graphics.scenery.utils.Wiggler
+import graphics.scenery.attribute.material.Material
 import graphics.scenery.utils.extensions.plus
 import graphics.scenery.utils.extensions.times
-import org.joml.Vector3f
-import java.lang.Thread.sleep
+import org.scijava.ui.behaviour.ClickBehaviour
 import kotlin.concurrent.thread
 import kotlin.system.exitProcess
 
 /**
  * Example for usage of VR controllers. Demonstrates the use of custom key bindings on the
- * HMD, the use of intersection testing with scene elements, and more advanced tools.
- *
- * Contents:
- * - Boxes to select and scale and to grab and move
- * - A grabable spray can to the left
- * - A touchable party sphere to the right
- *
- * Available Controls:
- * Side buttons alone:  Grab Object
- * Both side buttons together: Move to scale, after selection
- * Right Trigger:       Select to Scale
- * Left Trigger:        Select to Party first, then Scale
- * Left A Button:       Options Menu
+ * HMD, and the use of intersection testing with scene elements.
  *
  * @author Ulrik Günther <hello@ulrik.is>
  * @author Jan Tiemann <j.tiemann@hzdr.de>
  */
-class VRControllerExample : SceneryBase(
-    VRControllerExample::class.java.simpleName,
-    windowWidth = 1920, windowHeight = 1200
-) {
+class VRControllerExample : SceneryBase(VRControllerExample::class.java.simpleName,
+    windowWidth = 1920, windowHeight = 1200) {
     private lateinit var hmd: OpenVRHMD
     private lateinit var boxes: List<Node>
     private lateinit var hullbox: Box
-    private var leftControllerPushes = true
-    private var selectionStorage: Node? = null
 
     override fun init() {
         hmd = OpenVRHMD(useCompositor = true)
 
-        if (!hmd.initializedAndWorking()) {
+        if(!hmd.initializedAndWorking()) {
             logger.error("This demo is intended to show the use of OpenVR controllers, but no OpenVR-compatible HMD could be initialized.")
             exitProcess(1)
         }
@@ -60,15 +41,28 @@ class VRControllerExample : SceneryBase(
         renderer?.toggleVR()
 
         val cam: Camera = DetachedHeadCamera(hmd)
-        cam.spatial().position = Vector3f(0.0f, 0.0f, 0.0f)
+        cam.spatial {
+            position = Vector3f(0.0f, 0.0f, 0.0f)
+        }
 
         cam.perspectiveCamera(50.0f, windowWidth, windowHeight)
 
         scene.addChild(cam)
 
+        boxes = (0..10).map {
+            val obj = Box(Vector3f(0.1f, 0.1f, 0.1f))
+            obj.spatial {
+                position = Vector3f(-1.0f + (it + 1) * 0.2f, 1.0f, -0.5f)
+            }
+            obj
+        }
+
+        boxes.forEach { scene.addChild(it) }
+
+
         val lights = Light.createLightTetrahedron<PointLight>(spread = 5.0f, radius = 8.0f)
         lights.forEach {
-            it.emissionColor = Random.random3DVectorFromRange(0.8f, 1.0f)
+            it.emissionColor = Random.random3DVectorFromRange(0.0f, 1.0f)
             scene.addChild(it)
         }
 
@@ -82,126 +76,9 @@ class VRControllerExample : SceneryBase(
 
         scene.addChild(hullbox)
 
-        boxes = (0..10).map {
-            val obj = Box(Vector3f(0.1f, 0.1f, 0.1f))
-            obj.spatial {
-                position = Vector3f(-1.0f + (it + 1) * 0.2f, 1.0f, -0.5f)
-            }
-            obj.materialOrNull().diffuse = Vector3f(0.9f, 0.5f, 0.5f)
-            /**
-             * This attribute marks the box as touchable by [VRTouch].
-             * If there is an intersection with a box and the left controller,
-             * that box is slightly nudged in the direction
-             * of the controller's velocity.*/
-            obj.addAttribute(Touchable::class.java, Touchable(onTouch = { device ->
-                if (leftControllerPushes) {
-                    if (device.role == TrackerRole.LeftHand) {
-                        obj.ifSpatial {
-                            position = (device.velocity ?: Vector3f(0.0f)) * 0.05f + position
-                        }
-                    }
-                }
-            }))
-            obj.addAttribute(Grabable::class.java, Grabable())
-            obj.addAttribute(Selectable::class.java, Selectable(onSelect = { selectionStorage = obj }))
-            obj
-        }
-
-        boxes.forEach { scene.addChild(it) }
-
-        // scaled node - for grab testing
-        val scalePivot = RichNode().apply {
-            spatial {
-                scale = Vector3f(1/100f)
-                position = Vector3f(0.5f, 1.5f, 0f)
-            }
-            scene.addChild(this)
-        }
-        Sphere(10f).apply {
-            addAttribute(Grabable::class.java, Grabable())
-            addAttribute(Touchable::class.java, Touchable())
-            scalePivot.addChild(this)
-        }
-
-        // pressable sphere
-        val pressableSphere = Sphere(0.1f)
-        pressableSphere.spatial {
-            position = Vector3f(0.5f, 1.0f, 0f)
-        }
-        pressableSphere.addAttribute(
-            Pressable::class.java,
-            SimplePressable(onRelease = { Wiggler(pressableSphere.spatial(), 0.1f, 2000) })
-        )
-        scene.addChild(pressableSphere)
-
-        // remote controlled node
-        val rcBox = Box(Vector3f(0.1f,0.05f,0.07f)).apply {
-            spatial().position = Vector3f(0f, 1.0f, 1.0f)
-            scene.addChild(this)
-
-        }
-        // remote control
-        Sphere(0.05f).apply {
-            spatial().position = Vector3f(0f, 1.0f, 0.5f)
-            addAttribute(Grabable::class.java, Grabable(target = rcBox))
-            scene.addChild(this)
-
-        }
-
-        /** Box with rotated parent to debug grabbing
-        val pivot = RichNode()
-        pivot.spatial().rotation.rotateLocalY(Math.PI.toFloat())
-        scene.addChild(pivot)
-
-        val longBox = Box(Vector3f(0.1f, 0.2f, 0.1f))
-        longBox.spatial {
-        position = Vector3f(-0.5f, 1.0f, 0f)
-        }
-        longBox.addAttribute(Grabable::class.java, Grabable())
-        pivot.addChild(longBox)
-         */
-
-        // pen
-        val pen = Box(Vector3f(0.05f, 0.13f, 0.05f))
-        pen.spatial {
-            position = Vector3f(-0.5f, 1.0f, 0f)
-        }
-        scene.addChild(pen)
-        val tip = Box(Vector3f(0.025f, 0.025f, 0.025f))
-        tip.spatial {
-            position = Vector3f(0f, 0.08f, 0f)
-        }
-        pen.addChild(tip)
-        var lastPenWriting = 0L
-        pen.addAttribute(Grabable::class.java, Grabable())
-        pen.addAttribute(
-            Pressable::class.java, PerButtonPressable(
-                mapOf(
-                    OpenVRHMD.OpenVRButton.Trigger to SimplePressable(onHold = {
-                        if (System.currentTimeMillis() - lastPenWriting > 50) {
-                            val ink = Sphere(0.03f)
-                            ink.spatial().position = tip.spatial().worldPosition()
-                            scene.addChild(ink)
-                            lastPenWriting = System.currentTimeMillis()
-                        }
-                    }),
-                    OpenVRHMD.OpenVRButton.A to SimplePressable(onHold = {
-                        if (System.currentTimeMillis() - lastPenWriting > 50) {
-                            val ink = Box(Vector3f(0.03f))
-                            ink.spatial().position = tip.spatial().worldPosition()
-                            scene.addChild(ink)
-                            lastPenWriting = System.currentTimeMillis()
-                        }
-
-                    })
-                )
-            )
-        )
-
-
         thread {
             while (!running) {
-                sleep(200)
+                Thread.sleep(200)
             }
 
             hmd.events.onDeviceConnect.add { hmd, device, timestamp ->
@@ -237,125 +114,50 @@ class VRControllerExample : SceneryBase(
             }
         }
 
-        /** example of click input
         // Now we add another behaviour for toggling visibility of the boxes
         hmd.addBehaviour("toggle_boxes", ClickBehaviour { _, _ ->
-        boxes.forEach { it.visible = !it.visible }
-        logger.info("Boxes visible: ${boxes.first().visible}")
+            boxes.forEach { it.visible = !it.visible }
+            logger.info("Boxes visible: ${boxes.first().visible}")
         })
-        // ...and bind that to the A button of the left-hand controller.
-        hmd.addKeyBinding("toggle_boxes", TrackerRole.RightHand, OpenVRHMD.OpenVRButton.A)
-         */
+        // ...and bind that to the side button of the left-hand controller.
+        hmd.addKeyBinding("toggle_boxes", TrackerRole.LeftHand, OpenVRHMD.OpenVRButton.Side)
 
-        VRTouch.createAndSet(scene, hmd, listOf(TrackerRole.RightHand, TrackerRole.LeftHand), true)
+        // Finally, add a behaviour to toggle the scene's shell
+        hmd.addBehaviour("toggle_shell", ClickBehaviour { _, _ ->
+            hullbox.visible = !hullbox.visible
+            logger.info("Hull visible: ${hullbox.visible}")
+        })
+        //... and bind that to the A button on the left-hand controller.
+        hmd.addKeyBinding("toggle_shell", TrackerRole.LeftHand, OpenVRHMD.OpenVRButton.A)
 
-        VRGrab.createAndSet(
-            scene,
-            hmd,
-            listOf(OpenVRHMD.OpenVRButton.Side),
-            listOf(TrackerRole.LeftHand, TrackerRole.RightHand)
-        )
-        VRPress.createAndSet(
-            scene,
-            hmd,
-            listOf(OpenVRHMD.OpenVRButton.Trigger, OpenVRHMD.OpenVRButton.A),
-            listOf(TrackerRole.LeftHand, TrackerRole.RightHand)
-        )
+        // actions that need the position of the controller as an input need to be wrapped in the following lines
+        hmd.events.onDeviceConnect.add { hmd, device, timestamp ->
+            if(device.type == TrackedDeviceType.Controller) {
+                logger.info("Got device ${device.name} at $timestamp")
+                device.model?.let { controller ->
+                    // If a specific controller is desired they can be selected with
+                    // if (device.role == TrackerRole.LeftHand)
 
-        VRSelect.createAndSet(
-            scene,
-            hmd,
-            listOf(OpenVRHMD.OpenVRButton.Trigger),
-            listOf(TrackerRole.LeftHand),
-            { n ->
-                // this is just some action to show a successful selection.
-                // Party Cube!
-                val w = Wiggler(n.spatialOrNull()!!, 1.0f)
-                thread {
-                    sleep(2 * 1000)
-                    w.deativate()
+                    // This update routine is called every time the position of the controller
+                    // updates, and checks for intersections with any of the boxes. If there is
+                    // an intersection with a box, that box is slightly nudged in the direction
+                    // of the controller's velocity.
+                    controller.update.add {
+                        boxes.forEach { it.materialOrNull()?.diffuse = Vector3f(0.9f, 0.5f, 0.5f) }
+                        boxes.filter { box -> controller.children.first().spatialOrNull()?.intersects(box) ?: false }.forEach { box ->
+                            box.ifMaterial {
+                                diffuse = Vector3f(1.0f, 0.0f, 0.0f)
+                            }
+                            box.ifSpatial {
+                                position = (device.velocity ?: Vector3f(0.0f)) * 0.05f + position
+                            }
+                            (hmd as? OpenVRHMD)?.vibrate(device)
+                        }
+                    }
                 }
-            },
-            true
-        )
-
-        // a selection without a general action. Executes just the [onSelect] function of the [Selectable]
-        VRSelect.createAndSet(
-            scene,
-            hmd,
-            listOf(OpenVRHMD.OpenVRButton.Trigger),
-            listOf(TrackerRole.RightHand)
-        )
-
-        VRScale.createAndSet(hmd, OpenVRHMD.OpenVRButton.Side) {
-            selectionStorage?.ifSpatial { scale *= Vector3f(it) }
+            }
         }
-
-        val menu = VRSelectionWheel.createAndSet(
-            scene, hmd,
-            listOf(OpenVRHMD.OpenVRButton.Menu), listOf(TrackerRole.LeftHand),
-            listOf("Loading please wait" to {})
-        )
-        thread {
-            sleep(5000) // or actually load something
-            menu.get().actions = listOf(
-                "Toggle Shell" to {
-                    hullbox.visible = !hullbox.visible
-                    logger.info("Hull visible: ${hullbox.visible}")
-                },
-                "Toggle Boxes" to {
-                    boxes.forEach { it.visible = !it.visible }
-                    logger.info("Boxes visible: ${boxes.first().visible}")
-                },
-                "Tree Selection Wheel" to {
-                    val w = WheelMenu(hmd, listOf(
-                        Action("Test sub wheel") { println("test fix sub wheel") }
-                    ), true)
-                    w.spatial().position = menu.get().controller.worldPosition(Vector3f())
-                    scene.addChild(w)
-                },
-                "Toggle Push Left" to {
-                    leftControllerPushes = !leftControllerPushes
-                }).toActions()
-        }
-
-        VRTreeSelectionWheel.createAndSet(
-            scene, hmd,
-            listOf(OpenVRHMD.OpenVRButton.Menu), listOf(TrackerRole.RightHand),
-            listOf(
-                Switch("switch 1", false) { println("switch has been set to $it") },
-                Action("dummy1", false) { println("A dummy entry has been pressed") },
-                SubWheel(
-                    "Menu1", listOf(
-                        Action("dummy1-1") { println("A dummy entry has been pressed") },
-                        Action("dummy1-2") { println("A dummy entry has been pressed") },
-                        Action("dummy1-3") { println("A dummy entry has been pressed") }
-                    )
-                ),
-                SubWheel(
-                    "Menu2", listOf(
-                        Action("dummy2-1") { println("A dummy entry has been pressed") },
-                        SubWheel(
-                            "Menu2-1", listOf(
-                                Action("dummy2-1-1") { println("A dummy entry has been pressed") },
-                                Action("dummy2-1-2") { println("A dummy entry has been pressed") },
-                                Action("dummy2-1-3") { println("A dummy entry has been pressed") }
-                            )
-                        ),
-                        Action("dummy2-3") { println("A dummy entry has been pressed") }
-                    )
-                ),
-                Action("dummy2") { println("A dummy entry has been pressed") },
-                Action("dummy3") { println("A dummy entry has been pressed") },
-                Action("go to sleep") { thread {
-                    hmd.fadeToBlack()
-                    sleep(2000)
-                    hmd.fateToClear()
-                } },
-            )
-        )
     }
-
 
     companion object {
         @JvmStatic
